First test at using middleware as a processing pipeline.
I had thought of creating a plugable data pipeline for the pages, so
various processing could be done in a configurable way.  This is an
attempt to do this using http middleware instead of inventing another
pipeline system.

Currently some small portions of the page display are moved into the middleware.

Page lookup, parameter collection, and revision lookup are now done in middleware.

This will lead to changes in other api's as things are realigned to this workflow.